+++ NEWS.md
@@ -1,7 +1,11 @@
-# qlcVisualize 0.4 (upcoming)
+# qlcVisualize 0.4
 
-* adding option to regularize a weightedMap
-* catch erratic and unexplicable failing of concaveman
+* complete rewrite of weightedMap
+* catch erratic and unexplicable failing of concaveman in weightedMap
+* use two rounds of cartogram to prevent errors with complex maps
+* added second round of voronoi in weightedMap to regularize result
+* adding a worldwide polygon that can be used as a window for all glottolog languages
+* adding function addContour in preparation for rewrite of levelMap
 * various bugfixes
 
 # qlcVisualize 0.3